# docs/openapi.yaml
openapi: 3.0.3
info:
  title: パートナー協業プラットフォーム API
  description: |
    パートナー企業との協業を管理するためのREST API仕様です。

    ## 認証
    このAPIはBearer Token (JWT)認証を使用します。
    ログインAPIで取得したアクセストークンを`Authorization`ヘッダーに設定してください。

    ```
    Authorization: Bearer {accessToken}
    ```

    ## ページネーション
    一覧取得APIでは`limit`と`offset`パラメータでページネーションを制御できます。

    ## エラーレスポンス
    全てのエラーは統一フォーマットで返却されます。
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
servers:
  - url: https://api.example.com/api
    description: Production server
  - url: https://staging-api.example.com/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: ログイン
      description: ユーザー認証を行い、アクセストークンを取得します。
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: メールアドレスまたはパスワードが正しくありません
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: ログアウト
      description: 現在のセッションをログアウトします。
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '204':
          description: ログアウト成功
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: 現在のユーザー情報取得
      description: 現在ログイン中のユーザー情報を取得します。
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================
  # Project APIs
  # ============================================================
...
  /dashboard/my-today:
    get:
      tags:
        - Dashboard
      summary: 今日のタスク・アラート取得
      description: ログインユーザーの今日のタスクとアラートを取得します。
      operationId: getMyTodayDashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/MyTodayDashboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /dashboard/manager:
    get:
      tags:
        - Dashboard
      summary: マネージャーダッシュボード取得
      description: マネージャー向けダッシュボード情報を取得します。
      operationId: getManagerDashboard
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          description: 集計期間
          schema:
            type: string
            enum: [week, month, quarter]
            default: month
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ManagerDashboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ============================================================
  # Reminder APIs
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWTアクセストークン
...
  responses:
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: 認証が必要です
    Forbidden:
      description: 権限エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: アクセス権限がありません
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: エラーコード
            message:
              type: string
              description: エラーメッセージ
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: フィールド名
                  message:
                    type: string
                    description: 詳細メッセージ
...
    Priority:
      type: string
      enum: [low, medium, high]
      description: 優先度

    # ============================================================
    # Authentication Schemas
    # ============================================================
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: メールアドレス
          example: user@example.com
        password:
          type: string
          format: password
          description: パスワード
          example: password123
    LoginResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            accessToken:
              type: string
              description: アクセストークン
            refreshToken:
              type: string
              description: リフレッシュトークン
            expiresIn:
              type: integer
              description: トークン有効期限（秒）
            tokenType:
              type: string
              description: トークンタイプ
              example: Bearer
            user:
              $ref: '#/components/schemas/UserSummary'
    User:
      type: object
      properties:
        id:
          type: string
          description: ユーザーID
        email:
          type: string
          format: email
          description: メールアドレス
        name:
          type: string
          description: 氏名
        role:
          type: string
          enum: [admin, manager, member, partner]
          description: 役割
        department:
          type: string
          description: 部署
        partnerId:
          type: string
          nullable: true
          description: パートナーID
        permissions:
          type: array
          items:
            type: string
          description: 権限リスト
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
    UserSummary:
      type: object
      properties:
        id:
          type: string
          description: ユーザーID
        email:
          type: string
          format: email
          description: メールアドレス
        name:
          type: string
          description: 氏名
        role:
          type: string
          enum: [admin, manager, member, partner]
          description: 役割

    # ============================================================
    # Project Schemas
    # ============================================================
    ProjectStatus:
      type: string
      enum: [draft, active, completed, archived]
      description: 案件ステータス
...
    TaskStatus:
      type: string
      enum: [pending, in_progress, review, completed]
      description: タスクステータス
...
    MyTodayDashboard:
      type: object
      properties:
        date:
          type: string
          format: date
          description: 対象日
        user:
          $ref: '#/components/schemas/UserSummary'
        summary:
          type: object
          properties:
            totalTasks:
              type: integer
              description: 総タスク数
            completedTasks:
              type: integer
              description: 完了タスク数
            inProgressTasks:
              type: integer
              description: 進行中タスク数
            pendingTasks:
              type: integer
              description: 保留タスク数
            overdueTasks:
              type: integer
              description: 期限超過タスク数
        todayTasks:
          type: array
          items:
            $ref: '#/components/schemas/DashboardTask'
          description: 今日のタスク
        overdueTasks:
          type: array
          items:
            $ref: '#/components/schemas/OverdueTask'
          description: 期限超過タスク
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
          description: アラート
        upcomingMeetings:
          type: array
          items:
            $ref: '#/components/schemas/Meeting'
          description: 予定されている会議
    DashboardTask:
      type: object
      properties:
        id:
          type: string
          description: タスクID
        projectId:
          type: string
          description: 案件ID
        projectName:
          type: string
          description: 案件名
        title:
          type: string
          description: タスク名
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        dueDate:
          type: string
          format: date
          description: 期限日
        isOverdue:
          type: boolean
          description: 期限超過フラグ
    OverdueTask:
      allOf:
        - $ref: '#/components/schemas/DashboardTask'
        - type: object
          properties:
            daysOverdue:
              type: integer
              description: 超過日数
    Alert:
      type: object
      properties:
        id:
          type: string
          description: アラートID
        type:
          type: string
          enum: [task_overdue, project_milestone, project_deadline, budget_warning]
          description: アラートタイプ
        severity:
          type: string
          enum: [low, medium, high]
          description: 重要度
        message:
          type: string
          description: メッセージ
        relatedId:
          type: string
          description: 関連リソースID
        createdAt:
          type: string
          format: date-time
          description: 作成日時
    Meeting:
      type: object
      properties:
        id:
          type: string
          description: 会議ID
        title:
          type: string
          description: 会議名
        startTime:
          type: string
          format: date-time
          description: 開始時刻
        endTime:
          type: string
          format: date-time
          description: 終了時刻
        projectId:
          type: string
          description: 関連案件ID
        projectName:
          type: string
          description: 関連案件名
    ManagerDashboard:
      type: object
      properties:
        period:
          type: string
          enum: [week, month, quarter]
          description: 集計期間
        periodStart:
          type: string
          format: date
          description: 期間開始日
        periodEnd:
          type: string
          format: date
          description: 期間終了日
        projectSummary:
          type: object
          properties:
            total:
              type: integer
              description: 総案件数
            active:
              type: integer
              description: 進行中
            completed:
              type: integer
              description: 完了
            delayed:
              type: integer
              description: 遅延
            onTrack:
              type: integer
              description: 順調
            atRisk:
              type: integer
              description: リスクあり
        taskSummary:
          type: object
          properties:
            total:
              type: integer
              description: 総タスク数
            completed:
              type: integer
              description: 完了
            inProgress:
              type: integer
              description: 進行中
            pending:
              type: integer
              description: 保留
            overdue:
              type: integer
              description: 期限超過
            completionRate:
              type: number
              format: float
              description: 完了率（%）
        partnerPerformance:
          type: array
          items:
            $ref: '#/components/schemas/PartnerPerformance'
          description: パートナー実績
        projectsAtRisk:
          type: array
          items:
            $ref: '#/components/schemas/ProjectAtRisk'
          description: リスクのある案件
        recentActivities:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
          description: 最近のアクティビティ
        budgetOverview:
          $ref: '#/components/schemas/BudgetOverview'
        upcomingDeadlines:
          type: array
          items:
            $ref: '#/components/schemas/Deadline'
          description: 今後の期限
    PartnerPerformance:
      type: object
      properties:
        partnerId:
          type: string
          description: パートナーID
        partnerName:
          type: string
          description: パートナー名
        activeProjects:
          type: integer
          description: 進行中案件数
        tasksCompleted:
          type: integer
          description: 完了タスク数
        tasksTotal:
          type: integer
          description: 総タスク数
        onTimeDeliveryRate:
          type: number
          format: float
          description: 納期遵守率（%）
        rating:
          type: number
          format: float
          description: 評価
    ProjectAtRisk:
      type: object
      properties:
        id:
          type: string
          description: 案件ID
        name:
          type: string
          description: 案件名
        status:
          $ref: '#/components/schemas/ProjectStatus'
        progress:
          type: integer
          description: 進捗率
        daysRemaining:
          type: integer
          description: 残り日数
        overdueTaskCount:
          type: integer
          description: 期限超過タスク数
        riskLevel:
          type: string
          enum: [low, medium, high]
          description: リスクレベル
        riskReasons:
          type: array
          items:
            type: string
          description: リスク理由
    Activity:
      type: object
      properties:
        id:
          type: string
          description: アクティビティID
        type:
          type: string
          enum: [task_completed, task_created, project_created, project_updated, comment_added]
          description: アクティビティタイプ
        description:
          type: string
          description: 説明
        projectId:
          type: string
          description: 関連案件ID
        projectName:
          type: string
          description: 関連案件名
        userId:
          type: string
          description: ユーザーID
        userName:
          type: string
          description: ユーザー名
        createdAt:
          type: string
          format: date-time
          description: 発生日時
    BudgetOverview:
      type: object
      properties:
        totalBudget:
          type: integer
          description: 総予算（円）
        totalSpent:
          type: integer
          description: 総使用額（円）
        utilizationRate:
          type: number
          format: float
          description: 利用率（%）
        projectBudgets:
          type: array
          items:
            type: object
            properties:
              projectId:
                type: string
                description: 案件ID
              projectName:
                type: string
                description: 案件名
              budget:
                type: integer
                description: 予算（円）
              spent:
                type: integer
                description: 使用額（円）
              utilizationRate:
                type: number
                format: float
                description: 利用率（%）
    Deadline:
      type: object
      properties:
        type:
          type: string
          enum: [milestone, project, task]
          description: 期限タイプ
        projectId:
          type: string
          description: 関連案件ID
        projectName:
          type: string
          description: 関連案件名
        title:
          type: string
          description: タイトル
        dueDate:
          type: string
          format: date
          description: 期限日
        daysRemaining:
          type: integer
          description: 残り日数

    # ============================================================
    # Reminder Schemas
    # ============================================================
...